refactor: 메소드명에 포함된 "Str" 삭제, bonusBall 대신 lottoNumber를 사용하도록 수정
